LPS-130078 Adds useData hook to handle application data requests
useData is just a wrapper over useResource hook, the implementation just omits some properties that will be created in the data representation or because the application does not need those data stored in the state.

Some data is also relocated to allow the core components to continue to function, this may change in the future when everyone is consuming the same place for reducers from the same place.

When we are making the request we want to block the rendering, for reasons that we need this data in the components and that it is kept in the state, there are some problems:

- The rendering of the application can take a long time until the request is ready.

This can be smoothed out with the entry of `React.Suspense` but useResource is not compatible with that yet and has not left React's experimental state either.

Strategies like making requests earlier are also good ways to smooth the rendering of the application.